Refactor website structure for views
